asdasd

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: jheath <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/31 10:07:38 by jheath            #+#    #+#             */
-/*   Updated: 2018/06/12 11:08:35 by jheath           ###   ########.fr       */
+/*   Updated: 2018/08/02 12:00:21 by jheath           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int		ft_strlen(const char *s)
 {
 	unsigned int i;
